Store local checkout directory in autoloads indirection
The checkout directory is saved in a generated autoloads
indirection file when installing with
`package-vc-install-from-checkout'. the intention to use the
checkout directory in subsequent calls to `package-vc-update',
`package-vc-upgrade', and `package-vc-ugrade-all'.

* lisp/emacs-lisp/package-vc.el (package-vc--desc->spec):
Retrieve package spec from package's autoloads indirection file
when there's no spec in `package-vc-selected-packages'.
(package-vc--unpack-1): When installing from a local checkout,
store the value of `:lisp-dir' property of `pkg-spec' in a
"Package-spec" header in a generated autoloads indirection file.
(package-vc-upgrade): When package has a local checkout - as
defined by a presence of `:lisp-dir' property in `pkg-spec' -
use it as a place where package source has been checked out.

(Bug#79188)